Refactor the main module and split into separate modules
Refactor the main module and split it into separate modules for better organization and maintainability.

* **Add `image_processing.h` and `image_processing.cpp`**
  - Define types and declare functions for creating, initializing, displaying, and copying images.
  - Declare functions for image effects like transpose, mirror, fill, and negative filter.
  - Declare functions for color frequency analysis and sorting.
  - Declare functions for dynamic queue operations.
  - Implement the declared functions in `image_processing.cpp`.

* **Add `menu.h` and `menu.cpp`**
  - Declare functions for menu operations in `menu.h`.
  - Implement the declared functions in `menu.cpp`.

* **Modify `main.cpp`**
  - Include `menu.h` and `image_processing.h` header files.
  - Remove definitions of functions moved to `image_processing.cpp` and `menu.cpp`.
  - Update the `main` function to call the `Menu`, `menu`, and `handleMenuChoice` functions from `menu.h`.
